Basic fix for KeyTips not working when focus is inside a WinForms control
When pressing Alt and then a letter (input focus: WinForms textbox), Alt
will make the KeyTips show, but the letter will be typed into the
focused WinForms control instead of navigating through the ribbon. As
far as I can see, the reason for this is primarily that
OnFocusedElementPreviewTextInput can not be attached to the currently
focused WinForms control. This behaviour has not been an issue in older
versions of Fluent.

In order to fix the problem, in OnWindowKeyDown() e.Key will be
forwarded to the KeyTip adorners. This fix is incomplete though. Since
all keys must be set to "handled" here to avoid typing into the WinForms
textbox, this will also swallow the "up/down" navigation in menus. I'm
not quite sure what the best approach for fixing this should be.

I also modified the showcase app to show the two different textboxes for
testing the key tip behaviours in both scenarios.