Next try to change inheritance of DropDownButton
Inheritance of DropDownButton is changed from MenuBase to ItemsControl to overcome major focus, keyboard and usability problems. The inheritance from MenuBase especially hurts screen readers and people with disabilities.
All issues introduced for MenuItem by this change have to be solved in a way not involving inheritance from MenuBase.
First try for this change was made in 2014 on codeplex. Search for MenuBase in the issue tracker there to gather more information.